+++ server/common/src/middlewares/isUser.ts
@@ -1,10 +1,23 @@
 import { Request, Response, NextFunction } from 'express';
 import { NotAuthorizedError } from '../errors/notAuthorizedError';
+import { jwtPayload } from '../interfaces/payLoads';
+import { currentUser } from '../services/currentUser';
+
+declare global {
+  namespace Express {
+    interface Request {
+      currentUser?: jwtPayload;
+    }
+  }
+}
 
 export const isUser = (req: Request, res: Response, next: NextFunction) => {
-  if (req.currentUser?.isAdmin) {
-    throw new NotAuthorizedError();
-  }
+  const user = currentUser(req);
+  if (!user) throw new NotAuthorizedError();
+
+  if (user.isAdmin) throw new NotAuthorizedError();
+
+  req.currentUser = user;
 
   next();
 };
